refactor(gui): stage 2 — extract modal modules + registry seam
Verbatim moves out of main.js:
- src/app/modals/launcher.js — launcher UI/state + the
  duplicate/restart session actions that ride on it; focus callbacks
  (setFocusedTile, refocusActiveTerm) injected via initLauncher(deps).
- src/app/modals/project-editor.js — editor + its DOM listeners,
  bound in initProjectEditor(deps).
- src/app/modals/command-palette.js — palette UI only; the command
  TABLE stays in main.js (actions live there until later stages) and
  is handed over at initCommandPalette({commands}).
- src/app/modals/help-overlay.js — ⌘/ overlay + native-menu toggle.
- src/app/modals/registry.js — registerModal/anyModalOpen; the ONE
  intentional non-move edit: focusSnapshot's hard-coded four-element
  modal check becomes anyModalOpen() over the registered set.

main.js: 3420 → 2921 lines. Suites: vitest 149, mock e2e 52(+1 skip),
e2e-real 5 — all green.
